Changing duration returned from DIFF_IN_USEC to unsigned long
In the case that the diff returned from the start and end of the test
overflowed an unsigned int, making the container for the result
an unsigned long. (This is the type returned from the macro anyway).

+++ src/check_log.c
@@ -254,12 +254,12 @@ void xml_lfun (SRunner *sr CK_ATTRIBUTE_UNUSED, FILE *file, enum print_output pr
   case CLENDLOG_SR:
     {
       struct timespec ts_end;
-      unsigned int duration;
+      unsigned long duration;
 
       /* calculate time the test were running */
       clock_gettime(CLOCK_MONOTONIC, &ts_end);
-      duration = DIFF_IN_USEC(ts_start, ts_end);
-      fprintf(file, "  <duration>%u.%06u</duration>\n",
+      duration = (unsigned long)DIFF_IN_USEC(ts_start, ts_end);
+      fprintf(file, "  <duration>%lu.%06lu</duration>\n",
           duration / 1000000, duration % 1000000);
       fprintf(file, "</testsuites>\n");
     }
